feat(id-generator): use StrongUuidGenerator by default
The historic default DbIdGenerator is not safe to use in a cluster and should only be used in unit tests, but not in production. Since many users are using Spring Boot for microservices chances are very high that they have many applications using the same database and then run into deadlocks during db-based id generation.